Fixed some timing issue with client game and local entities. Decided to change how func_plat operates, q2 its way is ... ugly: 2 triggers, top and bottom, both can be disabled with a spawnflag. To have it start at top, another spawnflag has been introduced. Also it can be set to toggle mode so it won't automatically revert back.

+++ src/Server/Main.cpp
@@ -1766,7 +1766,7 @@ processing are run even when server is not yet initalized.
 Returns amount of extra frameTime available for sleeping on IO.
 ==================
 */
-unsigned SV_Frame(unsigned msec)
+uint64_t SV_Frame(uint64_t msec)
 {
 #if USE_CLIENT
     timeBeforeServerGame = timeAfterServerGame = 0;
